+++ pkg/analyzer/lib/dart/ast/ast.dart
@@ -189,8 +189,6 @@ export 'package:analyzer/src/dart/ast/ast.dart'
         NullAwareElement,
         NullCheckPattern,
         NullLiteral,
-        // ignore: deprecated_member_use_from_same_package
-        NullShortableExpression,
         ObjectPattern,
         ParenthesizedExpression,
         ParenthesizedPattern,