+++ pkg/analyzer/lib/dart/element/element2.dart
@@ -1,8 +1,0 @@
-// Copyright (c) 2024, the Dart project authors. Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-@Deprecated('Import element.dart instead')
-library;
-
-export 'package:analyzer/dart/element/element.dart';