+++ pkg/analyzer/lib/src/dart/element/member.dart
@@ -16,18 +16,6 @@ import 'package:analyzer/src/generated/utilities_dart.dart';
 import 'package:analyzer/src/utilities/extensions/element.dart';
 import 'package:pub_semver/pub_semver.dart';
 
-@Deprecated('Use SubstitutedExecutableElementImpl instead')
-typedef ExecutableMember = SubstitutedExecutableElementImpl;
-
-@Deprecated('Use SubstitutedFieldElementImpl instead')
-typedef FieldMember = SubstitutedFieldElementImpl;
-
-@Deprecated('Use SubstitutedElementImpl instead')
-typedef Member = SubstitutedElementImpl;
-
-@Deprecated('Use SubstitutedFormalParameterElementImpl instead')
-typedef ParameterMember = SubstitutedFormalParameterElementImpl;
-
 /// A constructor element defined in a parameterized type where the values of
 /// the type parameters are known.
 class SubstitutedConstructorElementImpl extends SubstitutedExecutableElementImpl
@@ -44,10 +32,6 @@ class SubstitutedConstructorElementImpl extends SubstitutedExecutableElementImpl
 
   @override
   InterfaceElementImpl get enclosingElement => baseElement.enclosingElement;
-
-  @Deprecated('Use enclosingElement instead')
-  @override
-  InterfaceElementImpl get enclosingElement2 => enclosingElement;
 
   @override
   ConstructorFragmentImpl get firstFragment => baseElement.firstFragment;
@@ -77,20 +61,10 @@ class SubstitutedConstructorElementImpl extends SubstitutedExecutableElementImpl
   @override
   LibraryElementImpl get library => baseElement.library;
 
-  @Deprecated('Use library instead')
-  @override
-  LibraryElementImpl get library2 => library;
-
   @override
   InternalConstructorElement? get redirectedConstructor {
     var element = baseElement.redirectedConstructor;
     return _redirect(element);
-  }
-
-  @Deprecated('Use redirectedConstructor instead')
-  @override
-  InternalConstructorElement? get redirectedConstructor2 {
-    return redirectedConstructor;
   }
 
   @override
@@ -107,20 +81,10 @@ class SubstitutedConstructorElementImpl extends SubstitutedExecutableElementImpl
     return _redirect(baseElement.superConstructor);
   }
 
-  @Deprecated('Use superConstructor instead')
-  @override
-  InternalConstructorElement? get superConstructor2 {
-    return superConstructor;
-  }
-
   @override
   T? accept<T>(ElementVisitor2<T> visitor) {
     return visitor.visitConstructorElement(this);
   }
-
-  @Deprecated('Use accept instead')
-  @override
-  T? accept2<T>(ElementVisitor2<T> visitor) => accept(visitor);
 
   @override
   void appendTo(ElementDisplayStringBuilder builder) {
@@ -212,16 +176,8 @@ abstract class SubstitutedElementImpl implements Element {
   @override
   String? get name => baseElement.name;
 
-  @Deprecated('Use name instead')
-  @override
-  String? get name3 => name;
-
   @override
   Element get nonSynthetic => baseElement;
-
-  @Deprecated('Use nonSynthetic instead')
-  @override
-  Element get nonSynthetic2 => nonSynthetic;
 
   @override
   AnalysisSession? get session => baseElement.session;
@@ -234,21 +190,9 @@ abstract class SubstitutedElementImpl implements Element {
     return baseElement.getExtendedDisplayName(shortName: shortName);
   }
 
-  @Deprecated('Use getExtendedDisplayName instead')
-  @override
-  String getExtendedDisplayName2({String? shortName}) {
-    return getExtendedDisplayName(shortName: shortName);
-  }
-
   @override
   bool isAccessibleIn(LibraryElement library) {
     return baseElement.isAccessibleIn(library);
-  }
-
-  @Deprecated('Use isAccessibleIn instead')
-  @override
-  bool isAccessibleIn2(LibraryElement library) {
-    return isAccessibleIn(library);
   }
 
   @override
@@ -260,21 +204,9 @@ abstract class SubstitutedElementImpl implements Element {
     return baseElement.thisOrAncestorMatching(predicate);
   }
 
-  @Deprecated('Use thisOrAncestorMatching instead')
-  @override
-  Element? thisOrAncestorMatching2(bool Function(Element e) predicate) {
-    return thisOrAncestorMatching(predicate);
-  }
-
   @override
   E? thisOrAncestorOfType<E extends Element>() {
     return baseElement.thisOrAncestorOfType<E>();
-  }
-
-  @Deprecated('Use thisOrAncestorOfType instead')
-  @override
-  E? thisOrAncestorOfType2<E extends Element>() {
-    return thisOrAncestorOfType();
   }
 
   @override
@@ -308,21 +240,11 @@ abstract class SubstitutedExecutableElementImpl extends SubstitutedElementImpl
     return [...typeParameters, ...formalParameters];
   }
 
-  @Deprecated('Use children instead')
-  @override
-  List<Element> get children2 {
-    return children;
-  }
-
   @override
   String? get documentationComment => baseElement.documentationComment;
 
   @override
   Element? get enclosingElement => baseElement.enclosingElement;
-
-  @Deprecated('Use enclosingElement instead')
-  @override
-  Element? get enclosingElement2 => enclosingElement;
 
   @override
   ExecutableFragmentImpl get firstFragment;
@@ -378,16 +300,8 @@ abstract class SubstitutedExecutableElementImpl extends SubstitutedElementImpl
   @override
   LibraryElement get library => baseElement.library;
 
-  @Deprecated('Use library instead')
-  @override
-  LibraryElement get library2 => library;
-
   @override
   MetadataImpl get metadata => baseElement.metadata;
-
-  @Deprecated('Use metadata instead')
-  @override
-  MetadataImpl get metadata2 => metadata;
 
   @override
   TypeImpl get returnType {
@@ -400,10 +314,6 @@ abstract class SubstitutedExecutableElementImpl extends SubstitutedElementImpl
   FunctionTypeImpl get type {
     return _type ??= substitution.mapFunctionType(baseElement.type);
   }
-
-  @Deprecated('Use typeParameters instead')
-  @override
-  List<TypeParameterElementImpl> get typeParameters2 => typeParameters;
 
   @override
   void appendTo(ElementDisplayStringBuilder builder) {
@@ -420,29 +330,11 @@ abstract class SubstitutedExecutableElementImpl extends SubstitutedElementImpl
     return builder.toString();
   }
 
-  @Deprecated('Use displayString instead')
-  @override
-  String displayString2({
-    bool multiline = false,
-    bool preferTypeAlias = false,
-  }) {
-    return displayString(
-      multiline: multiline,
-      preferTypeAlias: preferTypeAlias,
-    );
-  }
-
   @override
   void visitChildren<T>(ElementVisitor2<T> visitor) {
     for (var child in children) {
       child.accept(visitor);
     }
-  }
-
-  @Deprecated('Use visitChildren instead')
-  @override
-  void visitChildren2<T>(ElementVisitor2<T> visitor) {
-    return visitChildren(visitor);
   }
 
   static InternalExecutableElement from(
@@ -507,19 +399,11 @@ class SubstitutedFieldElementImpl extends SubstitutedVariableElementImpl
   @override
   List<Element> get children => const [];
 
-  @Deprecated('Use children instead')
-  @override
-  List<Element> get children2 => children;
-
   @override
   String? get documentationComment => baseElement.documentationComment;
 
   @override
   InstanceElement get enclosingElement => baseElement.enclosingElement;
-
-  @Deprecated('Use enclosingElement instead')
-  @override
-  InstanceElement get enclosingElement2 => enclosingElement;
 
   @override
   FieldFragmentImpl get firstFragment => baseElement.firstFragment;
@@ -539,12 +423,6 @@ class SubstitutedFieldElementImpl extends SubstitutedVariableElementImpl
       baseGetter,
       substitution,
     );
-  }
-
-  @Deprecated('Use getter instead')
-  @override
-  InternalGetterElement? get getter2 {
-    return getter;
   }
 
   @override
@@ -568,16 +446,8 @@ class SubstitutedFieldElementImpl extends SubstitutedVariableElementImpl
   @override
   LibraryElementImpl get library => baseElement.library;
 
-  @Deprecated('Use library instead')
-  @override
-  LibraryElement get library2 => library;
-
   @override
   MetadataImpl get metadata => baseElement.metadata;
-
-  @Deprecated('Use metadata instead')
-  @override
-  MetadataImpl get metadata2 => metadata;
 
   @override
   InternalSetterElement? get setter {
@@ -591,12 +461,6 @@ class SubstitutedFieldElementImpl extends SubstitutedVariableElementImpl
     );
   }
 
-  @Deprecated('Use setter instead')
-  @override
-  InternalSetterElement? get setter2 {
-    return setter;
-  }
-
   @override
   Version? get sinceSdkVersion => baseElement.sinceSdkVersion;
 
@@ -605,18 +469,8 @@ class SubstitutedFieldElementImpl extends SubstitutedVariableElementImpl
     return visitor.visitFieldElement(this);
   }
 
-  @Deprecated('Use accept instead')
-  @override
-  T? accept2<T>(ElementVisitor2<T> visitor) => accept(visitor);
-
   @override
   void visitChildren<T>(ElementVisitor2<T> visitor) {}
-
-  @Deprecated('Use visitChildren instead')
-  @override
-  void visitChildren2<T>(ElementVisitor2<T> visitor) {
-    visitChildren(visitor);
-  }
 
   static InternalFieldElement from(
     FieldElementImpl element,
@@ -670,12 +524,6 @@ class SubstitutedFieldFormalParameterElementImpl
     }
 
     return SubstitutedFieldElementImpl.from(field, substitution);
-  }
-
-  @Deprecated('Use field instead')
-  @override
-  FieldElement? get field2 {
-    return field;
   }
 
   @override
@@ -742,12 +590,6 @@ class SubstitutedFormalParameterElementImpl
     return [...typeParameters, ...formalParameters];
   }
 
-  @Deprecated('Use children instead')
-  @override
-  List<Element> get children2 {
-    return children;
-  }
-
   @override
   String? get defaultValueCode => baseElement.defaultValueCode;
 
@@ -756,10 +598,6 @@ class SubstitutedFormalParameterElementImpl
 
   @override
   Element? get enclosingElement => baseElement.enclosingElement;
-
-  @Deprecated('Use enclosingElement instead')
-  @override
-  Element? get enclosingElement2 => enclosingElement;
 
   @override
   FormalParameterFragmentImpl get firstFragment => baseElement.firstFragment;
@@ -812,16 +650,8 @@ class SubstitutedFormalParameterElementImpl
   @override
   LibraryElement? get library => baseElement.library;
 
-  @Deprecated('Use library instead')
-  @override
-  LibraryElement? get library2 => library;
-
   @override
   MetadataImpl get metadata => baseElement.metadata;
-
-  @Deprecated('Use metadata instead')
-  @override
-  MetadataImpl get metadata2 => metadata;
 
   @override
   String get nameShared => name!;
@@ -835,10 +665,6 @@ class SubstitutedFormalParameterElementImpl
   @override
   Version? get sinceSdkVersion => baseElement.sinceSdkVersion;
 
-  @Deprecated('Use typeParameters instead')
-  @override
-  List<TypeParameterElementImpl> get typeParameters2 => typeParameters;
-
   @override
   TypeImpl get typeShared => type;
 
@@ -846,10 +672,6 @@ class SubstitutedFormalParameterElementImpl
   T? accept<T>(ElementVisitor2<T> visitor) {
     return visitor.visitFormalParameterElement(this);
   }
-
-  @Deprecated('Use accept instead')
-  @override
-  T? accept2<T>(ElementVisitor2<T> visitor) => accept(visitor);
 
   @override
   void appendTo(ElementDisplayStringBuilder builder) {
@@ -859,12 +681,6 @@ class SubstitutedFormalParameterElementImpl
   @override
   void visitChildren<T>(ElementVisitor2<T> visitor) {
     baseElement.visitChildren(visitor);
-  }
-
-  @Deprecated('Use visitChildren instead')
-  @override
-  void visitChildren2<T>(ElementVisitor2<T> visitor) {
-    visitChildren(visitor);
   }
 
   static InternalFormalParameterElement from(
@@ -923,12 +739,6 @@ class SubstitutedGetterElementImpl
     );
   }
 
-  @Deprecated('Use correspondingSetter instead')
-  @override
-  InternalSetterElement? get correspondingSetter2 {
-    return correspondingSetter;
-  }
-
   @override
   GetterFragmentImpl get firstFragment => baseElement.firstFragment;
 
@@ -950,10 +760,6 @@ class SubstitutedGetterElementImpl
   T? accept<T>(ElementVisitor2<T> visitor) {
     return visitor.visitGetterElement(this);
   }
-
-  @Deprecated('Use accept instead')
-  @override
-  T? accept2<T>(ElementVisitor2<T> visitor) => accept(visitor);
 
   static InternalGetterElement forSubstitution(
     InternalGetterElement element,
@@ -1015,10 +821,6 @@ class SubstitutedMethodElementImpl extends SubstitutedExecutableElementImpl
   @override
   LibraryElement get library => baseElement.library;
 
-  @Deprecated('Use library instead')
-  @override
-  LibraryElement get library2 => library;
-
   @override
   Version? get sinceSdkVersion => baseElement.sinceSdkVersion;
 
@@ -1026,10 +828,6 @@ class SubstitutedMethodElementImpl extends SubstitutedExecutableElementImpl
   T? accept<T>(ElementVisitor2<T> visitor) {
     return visitor.visitMethodElement(this);
   }
-
-  @Deprecated('Use accept instead')
-  @override
-  T? accept2<T>(ElementVisitor2<T> visitor) => accept(visitor);
 
   static InternalMethodElement forTargetType(
     InternalMethodElement element,
@@ -1081,10 +879,6 @@ abstract class SubstitutedPropertyAccessorElementImpl
     return super.enclosingElement!;
   }
 
-  @Deprecated('Use enclosingElement instead')
-  @override
-  Element get enclosingElement2 => enclosingElement;
-
   @override
   PropertyAccessorFragmentImpl get firstFragment;
 
@@ -1103,12 +897,6 @@ abstract class SubstitutedPropertyAccessorElementImpl
       default:
         return variable;
     }
-  }
-
-  @Deprecated('Use variable instead')
-  @override
-  InternalPropertyInducingElement? get variable3 {
-    return variable;
   }
 
   @override
@@ -1146,12 +934,6 @@ class SubstitutedSetterElementImpl
     );
   }
 
-  @Deprecated('Use correspondingGetter instead')
-  @override
-  InternalGetterElement? get correspondingGetter2 {
-    return correspondingGetter;
-  }
-
   @override
   SetterFragmentImpl get firstFragment => baseElement.firstFragment;
 
@@ -1173,10 +955,6 @@ class SubstitutedSetterElementImpl
   T? accept<T>(ElementVisitor2<T> visitor) {
     return visitor.visitSetterElement(this);
   }
-
-  @Deprecated('Use accept instead')
-  @override
-  T? accept2<T>(ElementVisitor2<T> visitor) => accept(visitor);
 
   static InternalSetterElement forSubstitution(
     InternalSetterElement element,
@@ -1249,12 +1027,6 @@ class SubstitutedSuperFormalParameterElementImpl
       substitution,
     );
   }
-
-  @Deprecated('Use superConstructorParameter instead')
-  @override
-  InternalFormalParameterElement? get superConstructorParameter2 {
-    return superConstructorParameter;
-  }
 }
 
 /// A variable element defined in a parameterized type where the values of the
@@ -1314,18 +1086,6 @@ abstract class SubstitutedVariableElementImpl extends SubstitutedElementImpl
   @override
   String displayString({bool multiline = false, bool preferTypeAlias = false}) {
     return baseElement.displayString(
-      multiline: multiline,
-      preferTypeAlias: preferTypeAlias,
-    );
-  }
-
-  @Deprecated('Use displayString instead')
-  @override
-  String displayString2({
-    bool multiline = false,
-    bool preferTypeAlias = false,
-  }) {
-    return displayString(
       multiline: multiline,
       preferTypeAlias: preferTypeAlias,
     );